content/html/content/test/test_bug551670.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=551670
-->
<head>
  <title>Test for Bug 551670</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=551670">Mozilla Bug 551670</a>
<p id="display"></p>
<div id="content" style="display: none">
  <button type='badtype' id='b1'></button>
  <button id='b2'></button>
  <input type='badtype' id='i1'>
  <input id='i2'>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 551670 **/

function checkType(e1, e2, defaultType, otherType, wrongType)
{
  // When accessing to the type attribute from the IDL,
  // it should reflect the state of the button.

  defaultType = defaultType.toLowerCase();
  otherType = otherType.toLowerCase();

  is(e1.type, defaultType,
    "When the initial type attribute value is not valid, the state should be '" +
    defaultType + "'");
  is(e2.type, defaultType,
    "When the type attribute is not set, the state should be '" +
    defaultType + "'");

  e1.type = otherType;
  e1.setAttribute('type', '');
  is(e1.type, defaultType,
    "When type attribute is set to an empty string, the state should be '" +
    defaultType + "'");

  e1.type = otherType;
  e1.type = '';
  is(e1.type, defaultType,
    "When type attribute is set to an empty string, the state should be '" +
    defaultType + "'");

  e1.type = otherType;
  e1.setAttribute('type', wrongType);
  is(e1.type, defaultType,
    "When type attribute is set to an invalid value, the state should be '" +
    defaultType + "'");
  is(e1.getAttribute('type'), wrongType,
    "Type attribute content should not be changed");

  e1.type = otherType;
  e1.type = wrongType;
  is(e1.type, defaultType,
    "When type attribute is set to an invalid value, the state should be '" +
    defaultType + "'");
  is(e1.getAttribute('type'), wrongType,
    "Type attribute content should not be changed");

  e1.type = otherType.toUpperCase();
  is(e1.type, otherType, "Type attribute should be case insensitive");

  e1.removeAttribute('type');
  is(e1.type, defaultType,
    "When type attribute is set to an empty string, the state should be '" +
    defaultType + "'");
}

var wrongType = 'this-is-probably-a-wrong-type';

// button types
checkType(document.getElementById('b1'), document.getElementById('b2'), 'submit', 'button', wrongType);
checkType(document.getElementById('b1'), document.getElementById('b2'), 'submit', 'reset', wrongType);

// input types
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'button', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'checkbox', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'email', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'file', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'hidden', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'reset', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'image', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'password', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'radio', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'submit', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'tel', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'search', wrongType);
checkType(document.getElementById('i1'), document.getElementById('i2'), 'text', 'url', wrongType);

</script>
</pre>
</body>
</html>
